// front/src/app/model/user.model.ts
export class UserModel {
  id: string = '';
  first_name: string = '';
  last_name: string = '';
  email: string = '';
  password: string = ''
  role: string = '';
  token: string = '';
  avatar: string = '';
  description: string = '';
  language: string = '';
  location: string = '';
  title: string = '';
}